// src/styles/style.types.ts
export interface ITheme {
  fontPrimary: string;
  fontSecondary: string;
  border1: string;
  border2: string;
  backgroundPrimary: string;
  backgroundOpposite: string;
  boxShadowPrimary: string;
  boxShadowSecondary: string;
  buttonHoverBackground: string;
  buttonSetupProfilePicBackground: string;
  buttonConfirm: string;
  backgroundGlassNavbar: string;
  backgroundGlassNavbar2: string;
}